refactor: separate the colvarproxy_gpu and GPU helper functions
